Breaking out authorization specific views into their own file
