+++ website/views/__init__.py
@@ -1,2 +1,3 @@
+from .auth_views import *
 from .index_views import *
 from .product_details import *
